refactor(audit): split the store contracts from their implementations
Completes the previous commit, which carried only the renames.

store.ts keeps the AuditWriter/AuditBatchWriter/AuditReader contracts; each
backend lives in store.clickhouse.ts, store.drop.ts or store.pubsub.ts, so the
file list shows which store a deployment uses. store.ts is now type-only, so
importing DropAuditStore no longer pulls @clickhouse/client into the graph.

The 27 metadata shapes move out of event.ts, leaving the union and its
completeness check readable, and index.ts lists the 24 symbols consumers import
instead of re-exporting everything.

No behaviour change: excluding index.ts, which the narrowing rewrites, the
package is 519 code lines before and after with nothing on either side of the
diff.
