+++ packages/audit/lib/index.ts
@@ -1,8 +1,22 @@
-export type * from './event.js';
-export type { AuditBatchWriter, AuditReader, AuditWriter } from './store.js';
-export { ClickhouseAuditStore, DropAuditStore } from './store.js';
-export { PubSubAuditWriter } from './pubsub.js';
-export { AUDIT_DATABASE, auditClickhouseClient } from './clickhouse.js';
+export { auditClickhouseClient } from './clickhouse.js';
+export { AUDIT_EXPORT_MAX_ROWS, AuditClient, InvalidAuditCursorError } from './client.js';
 export { migrate } from './migrate.js';
-export { AUDIT_EXPORT_MAX_ROWS, AuditClient, InvalidAuditCursorError } from './audit.js';
-export { auditCsvHeader, auditCsvRows } from './csv.js';
+export { ClickhouseAuditStore } from './store.clickhouse.js';
+export { DropAuditStore } from './store.drop.js';
+export { PubSubAuditWriter } from './store.pubsub.js';
+export type {
+    AuditAction,
+    AuditActor,
+    AuditAttribution,
+    AuditContext,
+    AuditEvent,
+    AuditOutcome,
+    AuditResource,
+    AuditResourceAction,
+    AuditTarget,
+    AuditTargetType,
+    AuditVia,
+    NoAttribution
+} from './event.js';
+export type { AppAuthLoginMethod, MfaVerifiedMetadata } from './metadata.js';
+export type { AuditBatchWriter, AuditWriter } from './store.js';
